cannot overload __contains__ and return expr, resolves #31
`x in cpm_array([1,2,4])` calls `__contains__` on the NDVararray,
but whatever we return Python always converts it to `True` or `False`

XCSP3py overcomes this with a clever True/False detector and delayed
Expression generator. However, we can't do the same in CPMpy
as we have a more flexible, non-top-down expression generation process
that does not have a 'context' in which to do the delayed generation
e.g. `print(x in cpm_array([1,2,4]))` should return an Expression...